+++ server/entity/bulletin.go
