+++ server/entity/challenge.go
