+++ server/entity/replica.go
