+++ server/entity/replicaAlloc.go
