+++ server/services/database/resolvers/bulletin.go
